requirements.txt
stravalib
numpy
matplotlib
pyyaml
polyline
pandas
cartopy
scipy
plotly
